implement overflow protection
